Implement course page and part of lessons page

+++ src/main/java/com/university/dao/CourseRowMapper.java
@@ -11,7 +11,7 @@ public class CourseRowMapper implements RowMapper<Course> {
     public Course mapRow(ResultSet rs, int rowNum) throws SQLException {
         final Course course = new Course();
 
-        course.setId(rs.getInt("courseId"));
+        course.setId(rs.getLong("courseId"));
         course.setName(rs.getString("name"));
         course.setDescription(rs.getString("description"));
